Add files via upload
Bank is a class with fields account type and account balance. A parameterized constructor sets the account type and account balance. There is a method deposit that adds to the account balance and returns the value and another for withdrawal that subtracts from the account balance and returns the value. There is a display method that prints the account type and account balance.